restrucuring handler files
