+++ app/api/devices/handler.go
@@ -1,12 +1,11 @@
 package devices
 
 import (
-	"github.com/ahmadateya/flotta-webapp-backend/helpers"
-	"github.com/ahmadateya/flotta-webapp-backend/pkg/logparser"
-	"github.com/ahmadateya/flotta-webapp-backend/pkg/s3"
-	"github.com/gin-gonic/gin"
 	"net/http"
 	"strings"
+	"github.com/ahmadateya/flotta-webapp-backend/helpers"
+	"github.com/ahmadateya/flotta-webapp-backend/pkg/s3"
+	"github.com/gin-gonic/gin"
 )
 
 type Handler struct {
@@ -56,76 +55,6 @@ func (h *Handler) GetDeviceData(c *gin.Context) {
 	// return response
 	c.JSON(http.StatusOK, gin.H{
 		"data": "success",
-	})
-}
-
-func (h *Handler) GetNetworkData(c *gin.Context) {
-	device := c.Param("device")
-	// number of lines to read from log file
-	numOfLines := c.Query("logs")
-	if numOfLines == "" {
-		numOfLines = "2"
-	}
-
-	// read number of line n from the end of log file
-	raw, err := logparser.ReadLogFileRaw(device+"/network", numOfLines)
-	if err != nil {
-		helpers.FormatErrorMessage(c,
-			http.StatusInternalServerError,
-			"error reading network data",
-			err.Error(),
-		)
-		return
-	}
-	// parse raw log file into proper structs
-	netData, err := logparser.ParseNetworkRawLines(raw)
-	if err != nil {
-		helpers.FormatErrorMessage(c,
-			http.StatusInternalServerError,
-			"error parsing network data",
-			err.Error(),
-		)
-		return
-	}
-
-	// return response
-	c.JSON(http.StatusOK, gin.H{
-		"data": netData,
-	})
-}
-
-func (h *Handler) GetCPUTempData(c *gin.Context) {
-	device := c.Param("device")
-	// number of lines to read from log file
-	lines := c.Query("lines")
-	if lines == "" {
-		lines = "5"
-	}
-
-	// read number of line n from the end of log file
-	raw, err := logparser.ReadLogFileRaw(device+"/cputemp", lines)
-	if err != nil {
-		helpers.FormatErrorMessage(c,
-			http.StatusInternalServerError,
-			"error reading cpu temp data",
-			err.Error(),
-		)
-		return
-	}
-	// parse raw log file into proper structs
-	cpuData, err := logparser.ParseCPUTempRawLines(raw)
-	if err != nil {
-		helpers.FormatErrorMessage(c,
-			http.StatusInternalServerError,
-			"error parsing cpu temp data",
-			err.Error(),
-		)
-		return
-	}
-
-	// return response
-	c.JSON(http.StatusOK, gin.H{
-		"data": cpuData,
 	})
 }
